# packages/contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Salva Food API
  version: 0.1.0
servers:
  - url: /
security:
  - BearerAuth: []
paths:
  /api/v1/catalog/categories:
    get:
      summary: List catalog categories
      parameters:
      responses:
    Unauthorized:
      description: Missing, invalid, expired, or revoked session
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Authenticated principal lacks access or required permission
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
        '200':
          description: Categories ordered by sort_order
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
    post:
      summary: Create a catalog category
      parameters:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/catalog/items:
    get:
      summary: List catalog items
      parameters:
        - in: query
          name: category_id
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Items ordered by sort_order
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Item'
    post:
      summary: Create a catalog item
      parameters:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemRequest'
      responses:
        '201':
          description: Item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/orders:
    get:
      summary: List tenant orders
      parameters:
      responses:
        '200':
          description: Tenant orders
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
    post:
      summary: Create an order
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created or idempotently replayed
          headers:
            Idempotent-Replay:
              schema:
                type: string
                enum: ['true']
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/orders/{id}/transitions:
    post:
      summary: Transition an order state
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [status]
              properties:
                status:
                  $ref: '#/components/schemas/OrderStatus'
      responses:
        '200':
          description: Order transitioned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/auth/login:
    post:
      security: []
      summary: Authenticate a collaborator and create a session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/auth/logout:
    post:
      summary: Revoke the current session
      responses:
        '204': { description: Session revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/me:
    get:
      summary: Get the current authenticated principal
      responses:
        '200':
          description: Current principal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Principal' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/collaborators:
    get:
      summary: List tenant collaborators
      parameters:
        - in: query
          name: search
          schema: { type: string }
      responses:
        '200':
          description: Collaborators
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Collaborator' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      summary: Create a collaborator
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCollaboratorRequest' }
      responses:
        '201':
          description: Collaborator created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Collaborator' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/collaborators/{id}/status:
    patch:
      summary: Activate or deactivate a collaborator membership
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/MembershipStatus' }
      responses:
        '204': { description: Membership status updated }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 1
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Idempotency conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Category:
      type: object
      required: [tenant_id, id, name, sort_order, sold_out]
      properties:
        tenant_id: { type: string }
        id: { type: string }
        name: { type: string }
        sort_order: { type: integer, minimum: 0 }
        sold_out: { type: boolean }
    CreateCategoryRequest:
      type: object
      additionalProperties: false
      required: [name, sort_order]
      properties:
        name: { type: string, minLength: 1 }
        sort_order: { type: integer, minimum: 0 }
    Item:
      type: object
      required: [tenant_id, id, category_id, name, price_cents, sort_order, sold_out, sold_by_weight]
      properties:
        tenant_id: { type: string }
        id: { type: string }
        category_id: { type: string }
        name: { type: string }
        price_cents: { type: integer, format: int64, minimum: 0 }
        sort_order: { type: integer, minimum: 0 }
        sold_out: { type: boolean }
        sold_by_weight: { type: boolean }
    CreateItemRequest:
      type: object
      additionalProperties: false
      required: [category_id, name, price_cents, sort_order]
      properties:
        category_id: { type: string }
        name: { type: string, minLength: 1 }
        price_cents: { type: integer, format: int64, minimum: 0 }
        sort_order: { type: integer, minimum: 0 }
        sold_by_weight: { type: boolean, default: false }
    OrderStatus:
      type: string
      enum: [ANALYSIS, PRODUCTION, READY, FINALIZED, CANCELLED]
    OrderLine:
      type: object
      required: [item_id, name, quantity, unit_price_cents]
      properties:
        item_id: { type: string }
        name: { type: string }
        quantity: { type: integer, format: int64, minimum: 1 }
        unit_price_cents: { type: integer, format: int64, minimum: 0 }
    Order:
      type: object
      required: [tenant_id, id, source, status, items, total_cents]
      properties:
        tenant_id: { type: string }
        id: { type: string }
        source: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/OrderLine' }
        total_cents: { type: integer, format: int64, minimum: 0 }
    CreateOrderRequest:
      type: object
      additionalProperties: false
      required: [source, items]
      properties:
        source: { type: string, minLength: 1 }
        items:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required: [item_id, quantity]
            properties:
              item_id: { type: string }
              quantity: { type: integer, format: int64, minimum: 1 }
    MembershipStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
    Permission:
      type: string
      enum:
        - reports
        - virtual_assistant
        - sales_growth
        - subscription
        - settings
        - pos
        - establishment
        - coupon
        - satisfaction
        - menu_items
        - delivery_regions
        - buy_more
        - cashback
        - sales_recovery
        - hall_waiter
        - menu_edit
        - cash_open
        - ready_order_edit
        - fiscal_issue
        - nfce_issue
        - finance
        - purchasing
        - inventory
        - dashboard
        - online_payment
    Membership:
      type: object
      required: [tenant_id, user_id, title, status, permissions]
      properties:
        tenant_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        title: { type: string }
        status: { $ref: '#/components/schemas/MembershipStatus' }
        permissions:
          type: array
          items: { $ref: '#/components/schemas/Permission' }
    User:
      type: object
      required: [id, name, cpf, email, phone]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        cpf: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        image_url: { type: string }
    Principal:
      type: object
      required: [user, membership]
      properties:
        user: { $ref: '#/components/schemas/User' }
        membership: { $ref: '#/components/schemas/Membership' }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
        tenant_id: { type: string, format: uuid }
    LoginResponse:
      type: object
      required: [token, expires_at, principal]
      properties:
        token: { type: string }
        expires_at: { type: string, format: date-time }
        principal: { $ref: '#/components/schemas/Principal' }
    Collaborator:
      type: object
      required: [id, name, email, phone, title, status, permissions]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        image_url: { type: string }
        title: { type: string }
        status: { $ref: '#/components/schemas/MembershipStatus' }
        permissions:
          type: array
          items: { $ref: '#/components/schemas/Permission' }
    CreateCollaboratorRequest:
      type: object
      additionalProperties: false
      required: [name, cpf, email, password, phone, title, permissions]
      properties:
        name: { type: string, minLength: 1 }
        cpf: { type: string }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        phone: { type: string, minLength: 1 }
        title: { type: string, minLength: 1 }
        image_url: { type: string }
        permissions:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/Permission' }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message]
          properties:
            message: { type: string }